+++ test/instrumentation-tests/ppx-args.t
@@ -398,6 +398,7 @@ And a dune file:
   $ mutaml-runner _build/mutation/test.bc
   read mut file test.muts
   Testing without a mutant ... passed
+  The limit of a test run is 5 times the run without a mutant, and never less than 10 seconds.
   Testing mutant test.ml:l:int-constant:92be9951:0 ... passed
   Testing mutant test.ml:l:int-constant:50c5a5ff:0 ... passed
   Testing mutant test.ml:l:int-constant:626aba84:0 ... passed